requirements.txt
aiogram_dialog
pydantic
aio-pika
faststream
loguru
aiosqlite
pytz
apscheduler
fastapi
uvicorn
aiofiles
aiogram
aiohappyeyeballs
aiohttp
aiopg
aiosignal        
annotated-types
async-timeout
asyncpg
asyncpg-lite
attrs
certifi
frozenlist
greenlet
idna
magic-filter
multidict
pip
propcache
psycopg2-binary
python-dotenv
SQLAlchemy
typing_extensions
tzdata
tzlocal
yarl
pydantic-settings
alembic